Calculate features for epitope prediction

This function is used to calculate several distinct families of features for epitope prediction.

calc_features(
  peptides.list,
  local.features = character(),
  global.features = character(),
  ...
)

Arguments

peptides.list

list object returned by make_data_splits().

local.features

list of features to be calculated at the local neighbourhood (peptides.list$df$Info_window) level.

global.features

lists of features to be calculated at the global level (peptides.list$proteins$TSeq_sequence). See Feature Lists for details.

...

any other parameters (currently unused)

Value

Updated peptides.list object, with local features added as columns to peptides.list$df, and global features added as columns to peptides.list$proteins.

Details

Two major types of features can be calculated:

local features, which are calculated based on the local neighbourhood of each AA position (column Info_window of the windowed data frame).

global features, which are calculated using the full sequence of the protein (listed in column TSeq_sequence of the protein data frame).

The following features are calculated based on the implementations available in package protr. As of August 2021, the following groups of features are supported by the epitopes package (see the protr vignette for details on each of these):

"AAC" - Amino acid composition

"DC" - Dipeptide composition

"TC" - Tripeptide composition

CTD descriptors:

"CTDC" - Composition

"CTDT" - Transition

"CTDD" - Distribution

"CTriad" - Conjoint triad descriptors

Quasi-sequence-order descriptors:

"SOCN" - Sequence-order-coupling number (with maximum lag nlag = 3)

"QSO" - Quasi-sequence-order descriptors (with maximum lag nlag = 3 and weighting factor w = 0.1)

Proteochemometric Modeling descriptors:

"ScalesGap" - Scales-based descriptors derived by Principal Components Analysis (using all properties in the protr::AAindex matrix, pc = 5 and lag = 3)

NOTE: in all feature groups except "ScalesGap", invalid AA codes (B, J, O, U, X, Y) are removed from the strings prior to feature calculation. In "ScalesGap" these codes are replaced by a gap indicator, "-".

Besides those, the following features are also available based on native implementations:

"Entropy" - the Shannon entropy of a sequence.

"Atoms" - the number of C, H, N, O, S atoms in the sequence

"MolWeight" - the total molecular weight of the sequence

"AAtypes" - the proportion of AAs of each type (acidic, aliphatic, acidic, etc.)

"BLOSUM" - BLOSUM-derived descriptors (same as protr::extractBLOSUM() with submat = "AABLOSUM62", k = 5, lag = 3 and scale = TRUE)

"LegacyFeatures" - calculates the features used in paper doi: 10.1093/bioinformatics/btab536

Each feature group may be used either at the local or global level - which does not mean it should be. The protr documentation provides the following warning: "Users need to intelligently evaluate the underlying details of the descriptors provided, instead of using protr with their data blindly, especially for the descriptor types with more flexibility. It would be wise for the users to use some negative and positive control comparisons where relevant to help guide interpretation of the results.". Users should therefore be savvy when choosing which features to use for epitope prediction, and the choice should ideally be guided by domain expertise. Certain feature groups may not make sense as local features, as the (usually very short) local substrings will not allow the features to be informative (e.g., tripeptide composition, "TC"); or may require specific overriding of standard parameters (e.g., parameter lambda for "PAAC" must be smaller than the length of the shortest local string).

IMPORTANT: if the splitting level intended to be used for modelling is "peptide" (this can be checked in peptides.list$splits.attrs$split_level) then global features should be avoided, as their use can result in data leakage across splits and contaminate performance calculations.

Feature Vectors

Input vectors local.features and global.features are used to define which features are calculated at either level. These input parameters must be character vectors, where each element is one of the names listed in Details. For more information on the features calculated by protr, check ?protr::extractXYZ, replacing XYZ by the group abbreviation (see Details or the documentation of the protr package for the list of available feature groups). Notice that, for consistency purposes, the user parameters of all protr features are kept fixed in this routine. If the user wishes to add other features (or the same features with distinct parameters) they can calculate those separately and bind them to peptides.list$df (for local features) or peptides.list$proteins (for global ones). All feature columns should have names starting with "feat_local_" or "feat_global_".

Author

Felipe Campelo (f.campelo@aston.ac.uk)