Consolidate epitope and protein data

Label protein positions based on linear B-cell epitope information extracted from IEDB.

consolidate_data(
  epitopes,
  proteins,
  save_folder = NULL,
  only_exact = FALSE,
  set_positive = c("any", "mode", "all"),
  ncpus = 1
)

Arguments

epitopes

data frame of epitope data (returned by get_LBCE()).

proteins

data frame of protein data (returned by get_proteins()).

save_folder

path to folder for saving the results.

only_exact

logical, should only sequences labelled as "Exact Epitope" in variable epit_struc_def (within epitopes) be considered?

set_positive

how to decide whether an observation should be of the "Positive" (+1) or "Negative" (-1) class? Use "any" to set a sequence as positive if $n_positive > 0$, "mode" to set it if $n_positive >= n_negative$, or "all" to set it if $n_negative == 0$. Defaults to "mode".

ncpus

positive integer, number of cores to use

Value

Data frame containing all proteins mentioned that appear at least once in epitopes$protein_id (one row per aminoacid residue), containing the consolidated information extracted from epitopes.

Details

This routine is used to consolidate epitope information, by:

Filtering proteins such that only entries that appear at least once in epitopes$protein_id are retained

Using the epitopes data (fields protein_id, start_pos and end_pos) to generate labels for each position of the selected proteins.

Prior to the consolidation, entries in epitopes input are removed if they:

Lack a valid protein_id (i.e., if they have no corresponding entry in proteins)

Lack a valid string in epit_seq

Lack a valid definition in epit_struc_def

Have a mismatch between the sequence in epit_seq and the corresponding sequence between start_pos and end_pos on the protein sequence.

Entries in the proteins data frame with a value of TSeq_seqtype different from "protein" are also removed.

Author

Felipe Campelo (f.campelo@aston.ac.uk)