Filter epitopes by taxonomy and host

Filters an epitope dataframe by source organism and/or host IDs.

filter_epitopes(
  df,
  tax_list = NULL,
  orgIDs = NULL,
  hostIDs = NULL,
  removeIDs = NULL,
  orgID_column = "sourceOrg_id",
  hostID_column = "host_id"
)

Arguments

df

a data frame containing epitope data

tax_list

taxonomy list (generated by get_taxonomy())

orgIDs

vector of organism taxon IDs that we want to retain (using df$sourceOrg_id).

hostIDs

vector of host taxon IDs that we want to retain (using df$host_id).

removeIDs

vector of organism taxon IDs that we want to remove (using df$sourceOrg_id).

orgID_column

name of the column in df containing the source organism IDs

hostID_column

name of the column in df containing the host IDs

Value

data frame with epitope data, filtered by the criteria in orgIDs, hostIDs and removeIDs.

Author

Felipe Campelo (f.campelo@aston.ac.uk)