Fit Random Forest model to epitope data

Fits a Random Forest model to epitope data, using the data splits and features previously calculated using make_data_splits() and calc_features(), respectively.

fit_model(
  peptides.list,
  threshold = 0.5,
  holdout.split = NULL,
  CV.folds = NULL,
  sample.rebalancing = "undersampling",
  use.global.features = ifelse(peptides.list$splits.attrs$split_level == "protein",
    TRUE, FALSE),
  return.model = "full",
  ncpus = 1,
  rnd.seed = NULL,
  ...
)

Arguments

peptides.list

data frame containing the training data (one or more numerical predictors and one Class attribute).

threshold

probability threshold for attributing a prediction as positive.

holdout.split

name of split to be used as a holdout set. If NULL then the full data is used for training. Ignored if CV.folds is not NULL.

CV.folds

vector with the names of the splits to be used for cross validation. If NULL no cross-validation is performed.

sample.rebalancing

character: should the model try to compensate class imbalances during training? See Dealing with class imbalance for details.

use.global.features

logical: should global features (potentially available in peptides.list$proteins) be used? Should be left as the default unless the user knows exactly what they're doing. See Details.

return.model

model to be returned. Accepts "full" (return model trained on the full data); "partial" (model trained on all data except the holdout split; or on the set of data specified in all CV folds); or "none" (does not return a model).

ncpus

number of cores to use.

rnd.seed

seed for random number generator.

...

other options to be passed down to ranger::ranger().

Value

List containing the fitted model and several performance indicators.

Details

Function make_data_splits() defines data splits based on (protein or peptide) similarity. The split identifiers are stored in peptides.list$df$Info_split. Function calc_features() calculates the local and/or global features for each entry. Local features are stored in peptides.list$df as columns starting with feat_local_, whereas global features are stored in peptides.list$proteins, as columns starting with feat_global_. NOTE: Global features should only be used if the splitting level used was "protein", otherwise they may cause contamination of performance assessment due to data leakage. The splitting level of the data can be checked on peptides.list$splits.attrs$split_level.

Dealing with class imbalance

Parameter sample.rebalancing regulates how the modelling routine attempts to compensate class imbalances. The following strategies are available:

"undersampling" (default) : stratified undersampling of the majority class. The undersampling is performed in a manner that retains at least some observations related to each peptide from the majority class (plus all observations from the minority class).

"by_tree": sets the parameter case.weights of ranger::ranger() to give each observation a sampling probability inversely proportional to its class prevalence.

any other value: no rebalancing is done.

Performance assessment

This function has the following modes of performance assessment:

If both holdout.split and CV.folds are NULL, no performance assessment is done. A model is fit on the full data and returned.

If holdout.split is the valid name of a data split in peptides.list$df$Info_split AND CV.folds is NULL, a model is trained on the full data except the holdout.split and then applied to predict the labels for holdout.split. The performance returned corresponds to the performance on holdout.split.

If CV.folds is a vector of valid names of data splits in peptides.list$df$Info_split, the splits named are used as cross-validation folds. The performance returned corresponds to the average cross-validation performance using the CV.folds. If CV.folds is not NULL the value of holdout.split is ignored.

IMPORTANT: by default, the model returned by this routine is a model trained on the full data (fit after the performance is assessed on holdout splits or on cross-validation folds). This can be regulated by parameter return.model.

Author

Felipe Campelo (f.campelo@aston.ac.uk)