Extract linear B Cell epitopes from XML files retrieved from IEDB.

This function is used to extract information for linear B-cell epitopes from the XML files exported using the functionality provided by IEDB. It assumes that the user has downloaded the Complete Database Export from the XML Database Export field in IEDB's Database Export and extracted it in a given folder, which is passed as an argument to the function. This can be easily done with get_IEDB().

get_LBCE(data_folder, ncpus = 1, save_folder = NULL)

Arguments

data_folder

path (either relative or absolute) to the directory containing the XML files

ncpus

positive integer, number of cores to use

save_folder

path to folder saving the output.

Value

A data frame containing the epitope data.

Author

Felipe Campelo (f.campelo@aston.ac.uk)

Examples

my.dir   <- system.file("extdata/xml_examples", package="epitopes")
epitopes <- get_LBCE(my.dir)
#> Processing 37 files using 1 cores
#> Started at 2022-02-08 10:19:08
#> Ended at 2022-02-08 10:19:16
#> Elapsed time: 8.37 secs
#> Done!
#> 25 epitopes retrieved.
#> 0 processing errors.