Split epitope data.

Split the windowed epitope data returned by extract_peptides() into non-overlapping subsets. Proteins / peptides with similarities higher than a predefined threshold are always placed in the same split to prevent data leakage in machine learning. This routine tries to simultaneously approximate the user-defined proportions and maintain the overall class balance within each split.

make_data_splits(
  peptides.list,
  proteins,
  split_level = "protein",
  target_props = c(0.75, 0.25),
  split_names = NULL,
  similarity_threshold = 0.7,
  substitution_matrix = "BLOSUM62",
  alpha = 0.5,
  return_front = NULL,
  SAopts = list(),
  save_folder = NULL,
  ncpus = 1
)

Arguments

peptides.list

list object returned by extract_peptides(), containing the data frame of windowed epitope data (df) and the data frame of individual peptides (peptides).

proteins

data frame of proteins, returned by get_proteins() containing all proteins listed in peptides.list$df$Info_protein_id.

split_level

which sequences should be used to calculate similarity for splitting the data. Accepts "protein" (uses similarity of the full protein sequences, proteins$TSeq_sequence, to determine which observations should stay together in the splits) or "peptide" (uses similarity of the labelled peptides, peptides.list$peptides$Info_peptide). See Grouping strategy for details.

target_props

numeric vector of target proportions for each split (i.e., a vector (p1, p2, ..., pK) such that 0 < pk < 1 for all k and sum(pk) = 1).

split_names

optional, vector of names to be given to each split.

similarity_threshold

similarity threshold for grouping observations. See Grouping strategy for details.

substitution_matrix

character string indicating the substitution matrix to be used to calculate the peptide / protein alignments. Must be a matrix recognised by Biostrings::substitution.matrices() (e.g., "BLOSUM45", "PAM30", etc.)

alpha

weight parameter to regulate focus on maximising match to desired split proportions (target_props) vs. on approximating the class balance of the full data set. Must be a numeric value between 0 and 1. See Optimisation Problem for details.

return_front

should different tradeoff solutions (based on distinct values of alpha) be returned as well? Accepts either NULL (don't calculate front) or a positive integer greater than 2 (number of equally spaced alpha values between 0 and 1 to evaluate). If it's a positive integer, then the tradeoff dataset will be exported as object $splits.attrs$tradeoffs in the output list. Note that only non-dominated tradeoffs are returned, so the number of tradeoff points is likely to be smaller than the value of return_front. Usually 11 or 21 points are enough to give a good rough idea of the tradeoffs.

SAopts

list of control parameters to be used by the simulated annealing (SANN) algorithm. See stats::optim() for details.

save_folder

path to folder for saving the results. It will save the results as file peptides_list.rds (overwriting if necessary)

ncpus

positive integer, number of cores to use.

Value

A list object containing:

df: peptides.list$df updated to contain cluster number and split allocation of each entry

peptides: peptides.list$peptides updated to contain cluster number and split allocation of each entry

proteins: data frame containing data on all proteins listed in df$Info_protein_id.

peptide.attrs: list inherited from peptides.list, containing relevant attributes used in the earlier call to extract_peptides().

splits.attrs: list containing information about the splitting:

split_level: same as input parameter split_level

similarity_threshold: same as input parameter similarity_threshold

substitution_matrix: same as input parameter substitution_matrix

split_props: proportion of data allocated to each split

split_balance: proportion of positive observations in each split

target_props: same as input parameter target_props

target_balance: proportion of positive observations in full data

alpha: same as input parameter alpha

SW.scores: local alignment scores between each sequence (Smith-Waterman)

diss.matrix: dissimilarity matrix (see Grouping strategy for details)

clusters: hclust object with clustering structure.

cluster.alloc: data frame summarising the split allocations.

tradeoffs: data frame with tradeoffs between class balance and nominal split size as a function of alpha (only if return_front is not NULL).

If the splitting is impossible (e.g., if the number of clusters is smaller than the desired number of splits) the function throws a warning and returns a list with only SW.scores, diss.matrix and clusters.

Grouping strategy

The first step of this routine is to group the observations at either the proteins or the peptide level (depending on input parameter split_level). Local alignment scores for all pairs of sequences are calculated using the implementation of the Smith-Waterman algorithm available in function Biostrings::pairwiseAlignment(), with default parameters and the scoring matrix defined in substitution_matrix. These scores are returned as element SW.scores of the output list.

The dissimilarity matrix is calculated based on SW.scores, as:

diss[i,j] = 1 - SW.scores[i,j] / min(SW.scores[i,i], SW.scores[j,j])

which gives a value between 0 (perfect similarity) and 1 (maximum dissimilarity). diss[i,j] will be 0 if and only if the shorter sequence is fully and perfectly contained in the longer one; and will be 1 if and only if the Smith-Waterman alignment score is zero.

The dissimilarity matrix (returned as element diss.matrix of the output list) is used to calculate a hierarchical clustering of the sequences, and input parameter similarity_threshold is then used to define the resulting similarity clusters. Single-linkage is used, to guarantee that any pair of sequences with similarity greater than similarity_threshold will always be contained within the same cluster. The resulting clusters (returned as element clusters of the output list) represent the allocation units that are considered by the optimisation to split the data.

Optimisation Problem

This routine attempts to simultaneously minimise two objectives: (i) the difference between the actual proportion of data within each split and the desired levels informed by target_props, and (ii) the difference between the proportion of positive observations within each split and the overall proportion in the data. A simple linear aggregation strategy is used to define the following optimisation problem. Let:

nC: number of clusters.

K: number of splits.

xi: integer variable defining the split to which cluster i is allocated.

Ni+: number of positive observations in cluster i.

Ni: total number of observations in cluster i.

Gj*: desired proportion of data for split j.

P*: proportion of positive observations in the whole data.

We want to solve the problem:

minimize sum_j{ alpha x (Gj - Gj*)^2 + (1-alpha) x (pj - P*)^2 }

With:

xi \in {1, ..., K}, for all i = 1, ..., nC

yij = ifelse(xi == j, 1, 0)

Gj = sum_i{ yij * Ni } / sum_i{ Ni }

pj = sum_i{ yij * Ni+ } / sum_i{ yij * Ni }

Input parameter alpha regulates the relative importance attributed to each of the two objectives. At the limits, alpha = 0 ignores the desired proportions and focuses only on defining splits with class balances that are as close as possible to P*, whereas alpha = 0 ignores the class balance and tries to simply generate splits that are as faithful as possible to the desired proportions. An approximation to the Pareto-optimal front can be obtained by varying alpha between these two values.

The search space of this optimisation problem has a cardinality of K ^ nC. If the cardinality is under 10^6 possible allocations then this routine performs enumerative search and is guaranteed to return the global optimum. For larger search spaces a constructive heuristic followed by Simulated Annealing (see stats::optim() for details) is used. Input parameter SAopts can be used to pass control parameters to the Simulated Annealing.

Author

Felipe Campelo (f.campelo@aston.ac.uk)