Extract linear B Cell epitopes from XML files retrieved from IEDB.

This function is used to extract information for linear B-cell epitopes from the XML files exported using the functionality provided by IEDB. It assumes that the user has downloaded the Complete Database Export from the XML Database Export field in IEDB's Database Export and extracted it in a given folder, which is passed as an argument to the function. This can be easily done with get_IEDB().

get_LBCE(data_folder, ncpus = 1, save_folder = NULL)

Arguments

data_folder
path (either relative or absolute) to the directory containing the XML files

ncpus
positive integer, number of cores to use

save_folder
path to folder for saving the output.

Value

A data.table containing the epitope data is returned invisibly.

Author

Felipe Campelo (f.campelo@aston.ac.uk)

Examples

my.dir   <- system.file("extdata/xml_examples", package="epitopes")
epitopes <- get_LBCE(my.dir)

#> Processing 37 files using 1 cores 
#> Started at 2021-04-15 16:53:01 
#>   |>_______________________________________|3% processed. Elapsed time: 0.0 secs  |>>______________________________________|5% processed. Elapsed time: 0.1 secs  |>>>_____________________________________|8% processed. Elapsed time: 0.2 secs  |>>>>____________________________________|11% processed. Elapsed time: 0.4 secs  |>>>>>___________________________________|14% processed. Elapsed time: 0.6 secs  |>>>>>>__________________________________|16% processed. Elapsed time: 0.7 secs  |>>>>>>>_________________________________|19% processed. Elapsed time: 0.9 secs  |>>>>>>>>________________________________|22% processed. Elapsed time: 0.9 secs  |>>>>>>>>>_______________________________|24% processed. Elapsed time: 0.9 secs  |>>>>>>>>>>______________________________|27% processed. Elapsed time: 1.0 secs  |>>>>>>>>>>>_____________________________|30% processed. Elapsed time: 1.0 secs  |>>>>>>>>>>>>____________________________|32% processed. Elapsed time: 1.1 secs  |>>>>>>>>>>>>>___________________________|35% processed. Elapsed time: 1.5 secs  |>>>>>>>>>>>>>>__________________________|38% processed. Elapsed time: 1.5 secs  |>>>>>>>>>>>>>>>_________________________|41% processed. Elapsed time: 1.6 secs  |>>>>>>>>>>>>>>>>________________________|43% processed. Elapsed time: 2.1 secs  |>>>>>>>>>>>>>>>>>_______________________|46% processed. Elapsed time: 2.1 secs  |>>>>>>>>>>>>>>>>>>______________________|49% processed. Elapsed time: 2.2 secs  |>>>>>>>>>>>>>>>>>>>_____________________|51% processed. Elapsed time: 2.3 secs  |>>>>>>>>>>>>>>>>>>>>____________________|54% processed. Elapsed time: 2.4 secs  |>>>>>>>>>>>>>>>>>>>>>___________________|57% processed. Elapsed time: 2.5 secs  |>>>>>>>>>>>>>>>>>>>>>>__________________|59% processed. Elapsed time: 2.5 secs  |>>>>>>>>>>>>>>>>>>>>>>>_________________|62% processed. Elapsed time: 2.6 secs  |>>>>>>>>>>>>>>>>>>>>>>>>________________|65% processed. Elapsed time: 2.7 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>_______________|68% processed. Elapsed time: 2.8 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>______________|70% processed. Elapsed time: 3.8 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>_____________|73% processed. Elapsed time: 3.8 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>____________|76% processed. Elapsed time: 3.8 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>___________|78% processed. Elapsed time: 3.9 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>__________|81% processed. Elapsed time: 3.9 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>_________|84% processed. Elapsed time: 4.0 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>________|86% processed. Elapsed time: 4.0 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>_______|89% processed. Elapsed time: 4.2 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>______|92% processed. Elapsed time: 4.2 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>_____|95% processed. Elapsed time: 4.4 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>|100% processed. Elapsed time: 4.6 secs  |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>|100% processed. Elapsed time: 4.6 secsEnded at 2021-04-15 16:53:06 
#> Elapsed time: 4.69 secs
#> Done!
#>  25 epitopes retrieved.
#>  0 processing errors.